refactor : folder structure
