+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app';
 import { ChakraProvider } from '@chakra-ui/react';
-import Layout from '@components/Layout';
+import Layout from '@components/common/Layout';
 import { theme } from '@theme';
 
 function App({ Component, pageProps }: AppProps) {
